+++ devel/CHANGELOG.md
@@ -1,5 +1,9 @@
 # *Epeios* changelog
 
+2017-08-09 :
+- N4A, NJS, SCLN4A, SCLNJS :
+  - Moving *...NJS...* content to *...N4A...*, in an attempt to reuse the same libraries for all native components.
+
 2017-08-08 :
 - FLW :
   - Initializing some variables, although not necessary, to fix some warnings issued when compiling a *node.js* *addon* on installation on a *POSIX* OS.