+++ tools/esketch/CHANGELOG.md
@@ -1,5 +1,9 @@
 # *eSketch* changelog.
 
+2017-08-09 :
+- *esketch-node* :
+    - Using the *SCLN4A* library instead of *SCLNJS* one.
+
 2017-08-08 :
 - *esketch-node*
 	- Fixing error occurring when the path to the component contains a space.